requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.4.2
sqlalchemy==2.0.32
pydantic==2.8.2
face_recognition==1.3.0
